+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Unreleased
 ----------
-- Switched errors to `std::io::Error`
+- Switched errors to using `Box<dyn std::error::Error>`
   - Removed `anyhow` dependency
 - Removed deprecated `get_revision()` and `git_revision()` functions
 - Bumped minimum supported Rust version to `1.74`